+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v8.3.0 - ?
 
+- Added create_environment_reference plugin to create reference to environment variables
+- Added the 'value' attribute to std::testing::NullResource
 - Log warning when environment variable is not found by get_env plugin
 - Deprecate get_env_int plugin in favor of int(get_env())